authentication/requirements.txt
adium-theme-ubuntu==0.3.4
aniso8601==1.0.0
apt-xapian-index==0.46
beautifulsoup4==4.3.2
characteristic==0.1.0
chardet==2.3.0
command-not-found==0.3
DateTime==4.0.1
debtagshw==0.1
decorator==4.0.2
defer==1.0.6
dirspec==13.10
dj-database-url==0.3.0
Django==1.8.4
django-appconf==1.0.1
django-compressor==1.5
django-geoposition==0.2.2
djangorestframework==3.2.2
drf-nested-routers==0.9.0
duplicity==0.7.1
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-HTTPAuth==2.5.0
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
gyp==0.1
html5lib==0.999
httplib2==0.9
idna==0.9
infinity==1.3
intervals==0.3.1
itsdangerous==0.24
Jinja2==2.8
lockfile==0.8
lxml==3.4.2
MarkupSafe==0.23
marshmallow==1.2.6
oauthlib==0.6.1
oneconf==0.3.7.15.4.1
PAM==0.4.2
Pillow==2.7.0
piston-mini-client==0.7.5
py-bcrypt==0.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycrypto==2.6.1
pycups==1.9.72
pygobject==3.14.0
pyOpenSSL==0.13.1
pyserial==2.6
python-apt===0.9.3.11build1
python-bcrypt==0.3.1
python-debian==0.1.22
pytz==2015.4
pyxdg==0.25
service-identity==1.0.0
sessioninstaller==0.0.0
six==1.9.0
software-center-aptd-plugins==0.0.0
SQLAlchemy==1.0.8
SQLAlchemy-Utils==0.30.17
stevedore==1.3.0
system-service==0.1.6
Twisted-Core==14.0.2
Twisted-Web==14.0.2
unity-lens-photos==1.0
validators==0.8
vboxapi==1.0
virtualenv==13.1.1
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
WTForms-Alchemy==0.13.3
WTForms-Components==0.9.7
wxPython==3.0.1.1
wxPython-common==3.0.1.1
xdiagnose==3.7.2
zope.interface==4.1.1
